Add files via upload
##Introduction
This project focuses on cleaning and transforming data to ensure consistency and accuracy. The FIFA dataset provides a rich set of attributes about players but contains several inconsistencies in data formats. This includes non-standard formats in value columns, varied height and weight measurements, special characters, and mixed units that required cleaning. These inconsistencies could hinder analysis and need to be standardized to improve usability.
Therefore, the cleaning process focused on transforming the data to a standard format suitable for effective analysis.

##Objective
The aim of this project was to clean and transform a raw dataset of FIFA player information to ensure consistency, accuracy, and readiness for analysis.

##Dataset Summary
The dataset consisted of 18,980 rows and 75 columns, covering various attributes of FIFA players. Below is a step-by-step breakdown of the data cleaning process, including before-and-after snapshots of the dataset to illustrate the transformation process.

##Conclusion
The FIFA project showcased the importance of structured data cleaning and preparation in uncovering valuable insights. By leveraging Power Query functions in Excel, inconsistencies in the dataset were effectively resolved, resulting in a well-organized and reliable dataset. This process highlighted how essential data preprocessing is to ensure the accuracy and relevance of analysis.
The cleaned dataset can now be used for various exploratory and analytical purposes, such as performance benchmarking, player scouting, and trend analysis within the FIFA dataset. This project also underscored the role of Excel's powerful tools in handling real-world data challenges, solidifying its relevance in the field of data analysis.
As a next step, this dataset could be integrated with advanced visualization tools like Power BI or used for predictive modeling to gain deeper insights into player performance.